Move colless-like helpers into MDM file, keep node_depth helper with pandas
- Revert: keep _alifestd_calc_node_depth_asexual_contiguous in the
  pandas file (_alifestd_mark_node_depth_asexual.py); polars file
  imports from there
- Move _colless_like_fast_path, _colless_like_slow_path, and
  _alifestd_mark_colless_like_index_asexual_impl from the separate
  shared file into _alifestd_mark_colless_like_index_mdm_asexual.py
- Update var and sd variants to import from the MDM file
- Remove now-empty _alifestd_mark_colless_like_index_asexual.py
